提交网友"齐"更改的代码，修复编译告警信息。

+++ DuiVision/source/ControlBase.cpp
@@ -352,7 +352,7 @@ CControlBase* CControlBase::GetNextFocusableControl(CControlBase* pFocusControl)
 {
 	BOOL bStartSearch = FALSE;
 	// 先按照焦点控件查找一次
-	for (int i = 0; i < m_vecControl.size(); i++)
+	for (int i = 0; i < (int)m_vecControl.size(); i++)
 	{
 		CControlBase* pControlBase = m_vecControl.at(i);
 		if (pControlBase && pControlBase->GetVisible() && !pControlBase->GetDisable() && bStartSearch && pControlBase->IsTabStop())
@@ -370,7 +370,7 @@ CControlBase* CControlBase::GetNextFocusableControl(CControlBase* pFocusControl)
 	{
 		bStartSearch = TRUE;
 	}
-	for (int i = 0; i < m_vecControl.size(); i++)
+	for (int i = 0; i < (int)m_vecControl.size(); i++)
 	{
 		CControlBase* pControlBase = m_vecControl.at(i);
 		if(m_pControl == NULL)
